OrderedListEntry, UserChecklist, ChecklistManager models, specs; reorganize and refactor existing a bit.
