+++ db/migrate/20191130225826_create_ordered_list_entries.rb
@@ -4,7 +4,7 @@ class CreateOrderedListEntries < ActiveRecord::Migration[5.2]
     create_table :ordered_list_entries do |t|
       t.string :name, null: false
       t.string :description
-      t.integer :list_position, null: false, comment: 'This is zero-based. It is the order (positi.on) that this item should appear in its checklist'
+      t.integer :list_position, null: false, comment: 'This is zero-based. It is the order (position) that this item should appear in its checklist'
       t.string :ancestry
 
       t.timestamps
@@ -12,6 +12,5 @@ class CreateOrderedListEntries < ActiveRecord::Migration[5.2]
 
     add_index :ordered_list_entries, :ancestry
     add_index :ordered_list_entries, :name
-
   end
 end
